Adding test coverage

+++ src/main/java/com/tbutler78/minemapping/domain/County.java
@@ -10,8 +10,8 @@ import javax.persistence.*;
 @Table(name="county")
 public class County {
     @Id
-    @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private int id;
+    @GeneratedValue(strategy= GenerationType.IDENTITY)
+    private Long id;
     private String name;
 
     public County() {
@@ -22,13 +22,13 @@ public class County {
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
 
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
